Refactor to prepare for Python support
