+++ myprecious.gemspec
@@ -11,10 +11,13 @@ Gem::Specification.new do |s|
   s.add_runtime_dependency 'gems', '~> 1.0', '>= 1.0.0'
   s.add_runtime_dependency 'git', '~> 1.5', '>= 1.5.0'
   s.add_runtime_dependency 'rake-toolkit_program'
+  s.add_runtime_dependency 'rest-client', '~> 2.0.2', '>= 2.0'
+  s.add_runtime_dependency 'parslet', '~> 2.0.0'
   
   s.add_development_dependency "bundler", "~> 1.13"
   s.add_development_dependency "pry", "~> 0.13"
   s.add_development_dependency "rb-readline", "~> 0.5"
+  s.add_development_dependency "byebug"
   
   s.homepage    =
     'http://rubygems.org/gems/myprecious'
